Run Go tests in iq/ on c++ changes
iq/ mixes solutions in different languages in the same folder. Go does
not like unused C++ code in the same folder.

Current solution is to add an ignore go-build tag:

    // clang-format off-next-line
    //go:build ignore

This change ensures that Go tests run whenever C++ change is present to
verify it does not break go build.